Change `_nestElements` to dedupe HTML anchor tags for corner case usages
When you use jquery.tocify.js with a CMS WYSIWYG, anchor tags are saved
to the database, and then duplicated the next time the page loads. This
modification makes jquery.tocify.js only append HTML anchor tags to the
DOM if the id global attribute stored `hashValue` would be unique to
the whole document. This commit:

1. Modifies `_nestElements` to achieve the desired functionality.

2. Adds a new Jasmine test for this feature and updates existing tests.

3. Modifies demo files to demonstrate this enhancement.