Fix bug with links to internal anchors in other documents.
An anchor element such as `a[href=/foo#bar]` was not working
because the jQuery selector that assigns the scroll behavior
to internal anchor elements inside the `context` was matching
any anchor element with a hash in its `href` attribute. This
commit changes the behavior to only match anchor elements
where the `href` attribute starts with "#".